Refactor the download test helpers into the shared file.
