+++ server/webdriver/cdc_tests/autopush/download_test.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import tempfile
-
 from server.webdriver.cdc_tests.autopush.cdc_base_webdriver import \
     CdcAutopushTestBase
 from server.webdriver.shared_tests.download_test import DownloadTestMixin
@@ -21,20 +19,3 @@ from server.webdriver.shared_tests.download_test import DownloadTestMixin
 
 class TestDownload(DownloadTestMixin, CdcAutopushTestBase):
   """Class to test the download tool for Custom DC. Tests come from DownloadTestMixin"""
-
-  def setUp(self):
-    """
-        In addition to the base test setUp, need to also create a temporary
-        directory to use for downloaded files
-        """
-    self.downloads_folder = tempfile.TemporaryDirectory()
-    preferences = {"download.default_directory": self.downloads_folder.name}
-    super().setUp(preferences)
-
-  def tearDown(self):
-    """
-        In addition to base test tearDown, need to also clean up the temporary
-        directory that was created.
-        """
-    self.downloads_folder.cleanup()
-    super().tearDown()
